% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_nrows.R
\name{get_df_nrows}
\alias{get_df_nrows}
\title{Get a message with a number of rows}
\usage{
get_df_nrows(df)
}
\arguments{
\item{df}{Data.frame. The dataset to be used}
}
\value{
a message
}
\description{
Get a message with a number of rows
}
\examples{
get_df_nrows(df = iris)
get_df_nrows(df = mtcars)
}
